# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MiraiLink Backend API
  version: 2.0.0
  description: Contrato REST vigente de MiraiLink, incluyendo Afinidades y verificación de Google Play.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/chats/{chatId}/members:
    get:
      tags: [ Chats ]
      operationId: getChatMembers
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/ChatId'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/chats/private:
    post:
      tags: [ Chats ]
      operationId: createPrivateChat
      security: [ { bearerAuth: [] } ]
      requestBody: { $ref: '#/components/requestBodies/PrivateChat' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/chats/group:
    post:
      tags: [ Chats ]
      operationId: createGroupChat
      security: [ { bearerAuth: [] } ]
      requestBody: { $ref: '#/components/requestBodies/GroupChat' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Uuid: { type: string, format: uuid }
    Error:
      type: object
      required: [ code, message ]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        details:
          oneOf:
            - type: array
              items:
                type: object
            - type: object
              properties:
                capsule:
                  $ref: "#/components/schemas/CapsuleSnapshot"
...
    CapsuleSnapshot:
      type: object
      required:
        - id
        - userIds
        - status
        - progress
        - level
        - revision
      properties:
        id:
          type: string
          format: uuid
        userIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 2
          maxItems: 2
        status:
          type: string
          enum:
            - active
            - paused
            - left
            - cancelled
            - revealed
        progress:
          type: integer
          minimum: 0
          maximum: 4
        level:
          type: integer
          minimum: 0
          maximum: 4
        revision:
          type: integer
          minimum: 0
        rulesVersion:
          type: integer
        resumeAccepted:
          type: array
          items:
            type: string
            format: uuid
        revealRequestedBy:
          type:
            - string
            - "null"
          format: uuid
        question:
          type:
            - object
            - "null"
        completedQuestions:
          type: array
          items:
            type: object
            required:
              - questionId
              - instanceId
              - category
              - authorId
              - authorAnswer
              - peerAnswer
              - completedAt
            properties:
              questionId:
                type: string
              instanceId:
                type: string
                format: uuid
              category:
                type: string
              isCustom:
                type: boolean
              text:
                oneOf:
                  - type: string
                  - type: object
              authorId:
                type: string
                format: uuid
              authorAnswer:
                type: string
              peerAnswer:
                type: string
              completedAt:
                type: string
                format: date-time
      description: Snapshot autoritativo con histórico bilateral de preguntas completadas.
  parameters:
...
    ChatId:
      {
        name: chatId,
        in: path,
        required: true,
        schema: { $ref: '#/components/schemas/Uuid' }
      }
...
  requestBodies:
...
    PrivateChat:
      {
        required: true,
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [ otherUserId ],
                    properties: { otherUserId: { $ref: '#/components/schemas/Uuid' } }
                  }
              }
          }
      }
    GroupChat:
      {
        required: true,
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [ name, userIds ],
                    properties:
                      {
                        name: { type: string, minLength: 1, maxLength: 100 },
                        userIds:
                          {
                            type: array,
                            minItems: 1,
                            maxItems: 99,
                            items: { $ref: '#/components/schemas/Uuid' }
                          }
                      }
                  }
              }
          }
      }
...
  responses:
    Ok: { description: Operación correcta }
    Created: { description: Recurso creado }
    BadRequest:
      {
        description: Request inválido,
        content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
      }
...
    NotFound:
      {
        description: Recurso no encontrado,
        content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
      }
...